requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
azure-ai-inference==1.0.0b9
azure-core==1.34.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cmudict==1.0.32
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
Flask==3.1.1
google-auth==2.40.3
google-genai==1.19.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
language_tool_python==2.9.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
numpy==2.3.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
textstat==0.7.7
toml==0.10.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vaderSentiment==3.3.2
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
